% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{html_table}
\alias{html_table}
\title{Parse an html table into a list.}
\usage{
html_table(x, header = NA, trim = TRUE, fill = FALSE, dec = ".")
}
\arguments{
\item{x}{A node, node set or document.}

\item{header}{Use first row as header? If \code{NA}, will use first row
if it consists of \verb{<th>} tags.}

\item{trim}{Remove leading and trailing whitespace within each cell?}

\item{fill}{If \code{TRUE}, automatically fill rows with fewer than
the maximum number of columns with \code{NA}s.}

\item{dec}{The character used as decimal mark.}
}
\description{
Largely the same as \code{rvest::html_table()} with additional code to allow multiple lines & changed so parsed into a list.
}
\keyword{internal}
