% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_fetch.R
\name{rd_fetch}
\alias{rd_fetch}
\title{Fetch Data for a Species}
\usage{
rd_fetch(binomial)
}
\arguments{
\item{binomial}{\code{character}. The species's binomial name.}
}
\value{
A list of attributes and information about the species.
}
\description{
Pull data from the Reptiles Database for a specified species
}
\author{
Matt Lewis
}
