% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_species.R
\name{rd_species}
\alias{rd_species}
\title{Obtain basic info on all species present on the reptile database}
\usage{
rd_species(outdir)
}
\arguments{
\item{outdir}{(optional) \code{character} directory in which to save species list.
Defaults to a temporary file present only in memory.}
}
\value{
A \code{tibble}.
}
\description{
Downloads the most recent species checklist present on the reptile
database.
}
\author{
Matt Lewis
}
